refactor(mcp): split server into cohesive siblings
Decompose the MCP server into a curated directory module: service construction in mod.rs, tool execution in dispatch.rs, protocol adaptation in service.rs, stdio and hot-upgrade lifecycle in runtime.rs, and unit coverage in tests.rs.

Keep the published constructors and run_stdio exports stable, with no behavior changes to dormant mode, blocking dispatch, timeout policy, inflight accounting, or tool response shapes.

Fixes #916
