+++ crates/rag-rat-mcp/src/server/service.rs
@@ -1,0 +1,52 @@
+use rmcp::model::{
+    CallToolRequestParams, CallToolResult, Implementation, ListToolsResult, PaginatedRequestParams,
+    ServerCapabilities, ServerInfo, Tool,
+};
+use rmcp::service::RequestContext;
+use rmcp::{ErrorData, RoleServer, ServerHandler};
+use serde_json::{Map, Value};
+
+use super::RagRatService;
+
+impl ServerHandler for RagRatService {
+    fn get_info(&self) -> ServerInfo {
+        ServerInfo::new(ServerCapabilities::builder().enable_tools().build())
+            .with_server_info(Implementation::new("rag-rat", env!("CARGO_PKG_VERSION")))
+            .with_instructions(
+                "Read-only-source repo intelligence. Index and auto-heal writes are confined to \
+                 the configured SQLite database.",
+            )
+    }
+
+    async fn call_tool(
+        &self,
+        request: CallToolRequestParams,
+        _context: RequestContext<RoleServer>,
+    ) -> Result<CallToolResult, ErrorData> {
+        // The catalog (TOOL_NAMES / schema / description) is the single source of truth advertised
+        // by `list_tools`, and `call_tool_for_config` (via `call`) is the by-name dispatcher.
+        // Forward the raw request arguments straight to the `call()` chokepoint: no per-tool
+        // `#[tool]` forwarder and no `Parameters<T>` serialize->deserialize round-trip — the client
+        // JSON reaches exactly one deserialize, in `call_tool_with_db`.
+        let args = request.arguments.map(Value::Object).unwrap_or(Value::Null);
+        self.call_async(request.name.to_string(), args).await
+    }
+
+    async fn list_tools(
+        &self,
+        _request: Option<PaginatedRequestParams>,
+        _context: RequestContext<RoleServer>,
+    ) -> Result<ListToolsResult, ErrorData> {
+        let tools = crate::tools::TOOL_NAMES
+            .iter()
+            .map(|name| {
+                let input_schema = match crate::tools::schema(name) {
+                    Value::Object(map) => map,
+                    _ => Map::new(),
+                };
+                Tool::new((*name).to_string(), crate::tools::description(name), input_schema)
+            })
+            .collect();
+        Ok(ListToolsResult { tools, meta: None, next_cursor: None })
+    }
+}
